moved Follow implementations to accounts app
