+++ accounts/urls.py
@@ -8,4 +8,5 @@ urlpatterns = [
     path('signup/', views.SignUpView.as_view(), name='signup'),
     path('login/', auth_view.LoginView.as_view(template_name='accounts/login.html'), name='login'),
     path('logout/', auth_view.LogoutView.as_view(), name='logout'),
+    path('profile/<int:pk>/', views.UserDetailView.as_view(), name='profile')
 ]
