+++ accounts/views.py
@@ -1,8 +1,12 @@
 from django.shortcuts import redirect
-from django.contrib.auth.forms import UserCreationForm
-from django.contrib.auth import authenticate, login
-from django.views.generic.edit import FormView
 from django.urls import reverse_lazy
+from django.contrib.auth import authenticate, login
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth.models import User
+from django.views.generic.edit import FormView
+from django.views.generic.detail import DetailView
+from .models import Follows
 
 
 class SignUpView(FormView):
@@ -17,3 +21,19 @@ class SignUpView(FormView):
         user = authenticate(username=username, password=row_password)
         login(self.request, user)
         return redirect(self.get_success_url())
+
+
+class UserDetailView(LoginRequiredMixin, DetailView):
+    template_name = 'tmitt3r/profile.html'
+    model = User
+    context_object_name = 'profiled_user'
+    
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['follow_num'] = Follows.objects.filter(actor=self.object).count()
+        context['follower_num'] = Follows.objects.filter(followed_user=self.object).count()
+        if self.object != self.request.user:
+            context['is_others'] = True
+            if Follows.objects.filter(actor=self.request.user, followed_user=self.object).exists():
+                context['following'] = True
+        return context
