+++ tmitt3r/templates/tmitt3r/home.html
@@ -13,7 +13,7 @@
 <h1>Tmitt3r Home</h1>
 {% for tm33t in latest_tm33t_list %}
     <div>
-        <p>{{ tm33t.content }}</p>
+        <p>{{ tm33t.content|linebreaksbr }}</p>
         <p>Tm33ted : {{ tm33t.post_time }}</p>
         <p>By : @{{ tm33t.poster.get_username }}</p>
     </div>
