+++ tmitt3r/views.py
@@ -1,8 +1,9 @@
+from django.contrib.auth.models import User
 from django.shortcuts import render
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.views.generic.list import ListView
 from django.views.generic.edit import CreateView
-from django.utils import timezone
+from django.views.generic.detail import DetailView
 from django.urls import reverse_lazy
 from .models import Tm33t
